+++ src/main.c
@@ -8,8 +8,8 @@ int main(int argc, char ** argv){
 
     Matrix matrix = fileGetMatrix(argc, argv);
     
-    int dist = bioinsp(matrix, atoi((argv[3])), atoi((argv[4])), atoi((argv[5])));
-    printf("%d\n", dist);
+    float dist = bioinsp(matrix, atoi((argv[3])), atoi((argv[4])), atoi((argv[5])));
+    printf("%.2f\n", dist);
 
     deleteMatrix(&matrix);
